+++ MatrixKit/Models/MXKAppSettings.h
@@ -97,12 +97,12 @@
 @property (nonatomic) NSString *httpsLinkScheme;
 
 /**
- Indicate to hide un-decryptable events before joining the room. Default is `YES`.
+ Indicate to hide un-decryptable events before joining the room. Default is `NO`.
  */
 @property (nonatomic) BOOL hidePreJoinedUndecryptableEvents;
 
 /**
- Indicate to hide un-decryptable events in the room. Default is `YES`.
+ Indicate to hide un-decryptable events in the room. Default is `NO`.
  */
 @property (nonatomic) BOOL hideUndecryptableEvents;