+++ MatrixKit/Models/MXKAppSettings.m
@@ -108,8 +108,8 @@ static NSString *const kMXAppGroupID = @"group.org.matrix";
         {
             [[NSUserDefaults standardUserDefaults] setBool:YES forKey:@"sortRoomMembersUsingLastSeenTime"];
         }
-        _hidePreJoinedUndecryptableEvents = YES;
-        _hideUndecryptableEvents = YES;
+        _hidePreJoinedUndecryptableEvents = NO;
+        _hideUndecryptableEvents = NO;
         sortRoomMembersUsingLastSeenTime = YES;
         
         presenceColorForOnlineUser = [UIColor greenColor];