+++ CHANGES.rst
@@ -5,14 +5,13 @@ Features:
  * 
 
 Improvements:
- * 
+ * MXKAppSettings: Introduce `hideUndecryptableEvents`. Disabled by default.
 
 Bugfix:
  * 
 
 API Change:
- * Hide prejoined undecryptable events by default. See -[MXKAppSettings hidePreJoinedUndecryptableEvents].
- * Hide all undecryptable events by default. See -[MXKAppSettings hideUndecryptableEvents].
+ *
 
 Translations:
  *